requirements/requirements.txt
pytorch-lightning==1.6.4
wandb==0.12.19
pydantic==1.9.1
transformers[deepspeed]
zstandard==0.18.0
